requirements.txt
ultralytics==8.3.49
opencv-python-headless==4.10.0.84
paddlepaddle==3.0.0b1
paddleocr==2.9.1
paddle==
fastapi[standard]==0.115.6
uvicorn[standard]
gunicorn==23.0.0
pydantic==2.10.3
pymongo==4.9.2
motor==3.6.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
beanie==1.28.0
fastapi-mail==1.4.2
py-automapper==1.2.3